修改 oauth2 client 信息为mysql存储

+++ liuyun-auth/src/main/java/com/liuyun/auth/service/impl/AuthRedisCodeServiceImpl.java
@@ -1,14 +1,17 @@
 package com.liuyun.auth.service.impl;
 
 import com.liuyun.auth.service.AuthRedisCodeService;
+import com.liuyun.redis.constants.AuthRedisConstants;
 import com.liuyun.redis.service.RedisService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.oauth2.common.exceptions.InvalidGrantException;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.security.oauth2.provider.code.RandomValueAuthorizationCodeServices;
 import org.springframework.stereotype.Component;
 
+import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -39,14 +42,19 @@ public class AuthRedisCodeServiceImpl extends RandomValueAuthorizationCodeServic
      **/
     @Override
     protected void store(String code, OAuth2Authentication authentication) {
-        this.redisService.set(getKey(code), authentication, TIMEOUT, TimeUnit.SECONDS);
+        // 定制 redisTemplate value 序列化方式, 防止反序列化失败
+        RedisTemplate<String, Object> redisTemplate = redisService.getRedisTemplate2();
+        redisTemplate.opsForValue().set(getKey(code), authentication, TIMEOUT, TimeUnit.MINUTES);
     }
-
 
     @Override
     protected OAuth2Authentication remove(String code) {
         String codeKey = getKey(code);
-        OAuth2Authentication token = (OAuth2Authentication) redisService.get(codeKey);
+        Object obj = redisService.getRedisTemplate2().opsForValue().get(codeKey);
+        if (Objects.isNull(obj)) {
+            throw new InvalidGrantException("该 authorization_code 不存在 : " + code);
+        }
+        OAuth2Authentication token = (OAuth2Authentication) obj;
         this.redisService.delete(codeKey);
         return token;
     }
@@ -70,6 +78,6 @@ public class AuthRedisCodeServiceImpl extends RandomValueAuthorizationCodeServic
      * @date 2020/12/15 9:51 下午
      **/
     private String getKey(String code) {
-        return "oauth:code:" + code;
+        return AuthRedisConstants.getKey(AuthRedisConstants.AUTH_PREFIX, AuthRedisConstants.AUTHORIZATION_CODE_PREFIX, code);
     }
 }
